corpus/pyproject.toml
[project]
name = "corpus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain==0.3.0",
    "langchain-openai==0.3.1",
    "python-dotenv==1.0.1",
]
